created launcher file/dir structure
